精简合并项目，改部分pubilc， 升级电脑端的opentk和所有jint Texture2d重新支持mipmap（销锋镝铸修改）

+++ Engine/Engine.Graphics/FontBatch3D.cs
@@ -24,7 +24,7 @@ namespace Engine.Graphics
 				{
 					case '\n':
 						num++;
-						v = vector2 + ((float)num * (base.Font.GlyphHeight + base.Font.Spacing.Y + spacing.Y) * down);
+						v = vector2 + (num * (base.Font.GlyphHeight + base.Font.Spacing.Y + spacing.Y) * down);
 						continue;
 					case '\r':
 						continue;
